Refactor & Code Separation
Refactored common code for re-usability and enabled both web socket
server (socket.io) and tcp-server (net)
