+++ src/app/core/guid.js
